Move bundle comparison pages and update routes
