+++ src/app/bundles/ui/livog-ui/vs/shadcn/page.tsx
@@ -1,14 +1,24 @@
-import { ImpactTable } from '@/components/impact-table'
+// This page imports all Livog UI component examples for bundle size comparison
+'use client'
 
-export default function Page() {
+import { AccordionPage } from '@/app/bundles/ui/livogui/accordion/client'
+import { CheckboxPage } from '@/app/bundles/ui/livogui/checkbox/client'
+import { CollapsiblePage } from '@/app/bundles/ui/livogui/collapsible/client'
+import { DialogPage } from '@/app/bundles/ui/livogui/dialog/client'
+import { PopoverPage } from '@/app/bundles/ui/livogui/popover/client'
+import { SwitchPage } from '@/app/bundles/ui/livogui/switch/client'
+import { TabsPage } from '@/app/bundles/ui/livogui/tabs/client'
+
+export default function LivoguiBundlePage() {
   return (
-    <div className="mx-auto my-8 max-w-2xl">
-      <h1 className="mb-4 text-2xl font-semibold">Livog UI vs ShadCN</h1>
-      <ImpactTable
-        routePattern="/bundles/ui-bundle/:name"
-        baselineRoute="/bundles/ui-bundle/shadcn"
-        trimPrefix={["/bundles/ui-bundle/"]}
-      />
+    <div className="hidden">
+      <AccordionPage />
+      <CheckboxPage />
+      <CollapsiblePage />
+      <DialogPage />
+      <PopoverPage />
+      <SwitchPage />
+      <TabsPage />
     </div>
   )
 }
