+++ src/app/bundles/ui/shadcn/vs/livog-ui/page.tsx
@@ -1,14 +1,28 @@
-import { ImpactTable } from '@/components/impact-table'
+// This page imports matching ShadCN components for bundle size comparison
+'use client'
 
-export default function Page() {
+import { AccordionPage } from '@/app/bundles/ui/shadcn/accordion/client'
+import { DialogPage } from '@/app/bundles/ui/shadcn/dialog/client'
+import { AlertDialogPage } from '@/app/bundles/ui/shadcn/alert-dialog/client'
+import { SheetPage } from '@/app/bundles/ui/shadcn/sheet/client'
+import { CheckboxPage } from '@/app/bundles/ui/shadcn/checkbox/client'
+import { CollapsiblePage } from '@/app/bundles/ui/shadcn/collapsible/client'
+import { PopoverPage } from '@/app/bundles/ui/shadcn/popover/client'
+import { SwitchPage } from '@/app/bundles/ui/shadcn/switch/client'
+import { TabsPage } from '@/app/bundles/ui/shadcn/tabs/client'
+
+export default function ShadcnBundlePage() {
   return (
-    <div className="mx-auto my-8 max-w-2xl">
-      <h1 className="mb-4 text-2xl font-semibold">ShadCN vs Livog UI</h1>
-      <ImpactTable
-        routePattern="/bundles/ui-bundle/:name"
-        baselineRoute="/bundles/ui-bundle/livogui"
-        trimPrefix={["/bundles/ui-bundle/"]}
-      />
+    <div className="hidden">
+      <AccordionPage />
+      <AlertDialogPage />
+      <DialogPage />
+      <SheetPage />
+      <CheckboxPage />
+      <CollapsiblePage />
+      <PopoverPage />
+      <SwitchPage />
+      <TabsPage />
     </div>
   )
 }
